refactor interest area service
